Changed panic with warn in delete function and managed Bindings.conf

+++ components/script/dom/webglbuffer.rs
@@ -79,7 +79,10 @@ impl DroppableWebGLBuffer {
     }
 
     pub(crate) fn delete(&self, operation_fallibility: Operation) {
-        assert!(self.is_deleted());
+        if !self.is_deleted() {
+            warn!("DroppableWebGLBuffer::delete called on a buffer that is not deleted");
+            return;
+        }
         let context = self.context.root().unwrap();
         let cmd = WebGLCommand::DeleteBuffer(self.id);
         match operation_fallibility {
